Moved base out of plugin.rb
